1、解决数据插值问题 2、解决kafka客户端启动问题 3、解决部分汉化问题

+++ kafka-eagle-web/src/main/java/org/smartloli/kafka/eagle/web/service/OffLineLearningService.java
@@ -88,7 +88,7 @@ public class OffLineLearningService {
                         metricName[i] = queryResult.getMetric();
                         List<Double> doubles = new ArrayList<>();
                         dataLengthArr[u] = queryResult.getDps().size();
-                        for (int key : queryResult.getDps().keySet()) {
+                        for (long key : queryResult.getDps().keySet()) {
                             doubles.add(queryResult.getDps().get(key).doubleValue());
                         }
                         dataPoint.add(doubles);
